Adicionado problema 2;

+++ ProblemsSolver/Problem1Solver.cs
@@ -2,9 +2,9 @@ namespace ProblemsSolver
 {
     public class Problem1Solver
     {
-        public decimal Solve()
+        public int Solve()
         {
-            decimal result = 0;
+            int result = 0;
             for (int i = 0; i < 1000; i++)
             {
                 if (i%3 == 0 || i%5 == 0)
